Added :
Checking if the user is currently logged in, when they click on the login button, if they are a logout button is presented with the current user information.
Autmatic logout within 10minutes.
Removed the Delete/Edit buttons for unauthenticated users.
Re-styled the quotes page by adding a quote carousel which goes through the quotes and displays them at the top of the page.
Added an alert dialog to confirm the deletion of selected quotes.